move wopi middleware and app middleware into separate folders
