+++ files/test.wopitest
@@ -1,1 +1,1 @@
-WordBlankDocumentResources\WordBlankDocument.docxWordBlankDocument.docx
+WordComplexDocumentResources\WordComplexDocument.docxWordComplexDocument.docx
