+++ src/middleware/appMiddleware/createEmptyFile.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from 'express';
 import { writeFile } from 'fs/promises';
-import { fileInfo } from '../utils';
+import { fileInfo } from '../../utils';
 
 export async function createEmptyFile(req:Request, res: Response, next: NextFunction) {
   const { file_id: fileId } = req.params;
