+++ src/middleware/appMiddleware/getDiscoveryInfo.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from 'express';
-import { getWopiMethods } from '../utils';
+import { getWopiMethods } from '../../utils';
 
 export async function getDiscoveryInfo(req: Request, res: Response, next: NextFunction): Promise<void> {
   const data = await getWopiMethods();
