+++ src/middleware/wopiMiddleware/index.ts
@@ -1,13 +1,7 @@
-export * from './checkAccess';
 export * from './checkFileInfo';
-export * from './copyFile';
-export * from './createEmptyFile';
 export * from './deleteFile';
-export * from './getDiscoveryInfo';
 export * from './getFile';
-export * from './getFileNames';
 export * from './getLock';
-export * from './getRawBody';
 export * from './handleHeaders';
 export * from './lock';
 export * from './putFile';
