+++ src/middleware/wopiMiddleware/lock.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from 'express';
-import { fileInfo } from '../utils';
+import { fileInfo } from '../../utils';
 
-// const { fileInfo } = require('../utils');
+// const { fileInfo } = require('../../utils');
 
 export function lock(req: Request, res: Response, next: NextFunction): void {
   const lockValue = req.header('X-WOPI-Lock');
